feat(planner): Add "Bun" support
The installation and build logic are basically the same
as Node.js, so we reuse the Node.js planner but with
a special "Bun" flag to indicate that we are facing
a Bun project.

Don't repeat ourselves; however I may consider refactor
this part when Bun's ecosystem is being complete.