refactor(githubfs): Use ZipFS
The TAR retrieved from GitHub may be corrupted. For example, I can see there is a package.json file, but I cannot read it. Therefore, I switched to zipfs, which might be more stable.

This commit also adds a new error message: "repo is too large," since we read the stream to the RAM now. Note that it may introduce $2n$ memory usage.